If we're trying to get an Integer from a string literal, generate an Integer instead of a Long. Should fix #26.

+++ server/src/test/java/com/readytalk/revori/util/UtilTest.java
@@ -86,7 +86,7 @@ public class UtilTest {
 		
 		Object output = Util.convert(Integer.class, input);
 
-		long result = (Long) output;
+		int result = (Integer) output;
 		assertEquals(value, result);
 	}
 }
